refactor: 카테고리 이미지 파일 분리
