+++ src/constants/images.ts
@@ -1,0 +1,15 @@
+import koreanCategoryIcon from "../../assets/category-korean.png";
+import chineseCategoryIcon from "../../assets/category-chinese.png";
+import japaneseCategoryIcon from "../../assets/category-japanese.png";
+import asianCategoryIcon from "../../assets/category-asian.png";
+import westernCategoryIcon from "../../assets/category-western.png";
+import etcCategoryIcon from "../../assets/category-etc.png";
+
+export const RESTAURANT_IMAGE = {
+  한식: koreanCategoryIcon,
+  중식: chineseCategoryIcon,
+  일식: japaneseCategoryIcon,
+  아시안: asianCategoryIcon,
+  양식: westernCategoryIcon,
+  기타: etcCategoryIcon,
+};
